requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework-simplejwt==5.5.1
